+++ src/runtime/plugins/InputComponent.ts
@@ -1,12 +1,13 @@
+import { defineNuxtPlugin } from '#imports'
 export default defineNuxtPlugin((nuxtApp) => {
-  nuxtApp.hook("app:beforeMount", () => {
-    window.addEventListener("click", (event:any) => {
-        if (!event.target.matches(".glorious-input-field")) {
-            const options = document.querySelectorAll(".glorious-input-options");
-            options.forEach((el) => {
-              el.classList.add("hidden");
-            });
-          }
-    });
-  });
-});
+  nuxtApp.hook('app:beforeMount', () => {
+    window.addEventListener('click', (event: any) => {
+      if (!event.target.matches('.glorious-input-field')) {
+        const options = document.querySelectorAll('.glorious-input-options')
+        options.forEach((el) => {
+          el.classList.add('hidden')
+        })
+      }
+    })
+  })
+})